requirements.txt
pyyaml==5.3
SQLAlchemy==1.3.12
mysqlclient==1.4.6
requests==2.22.0
python-dateutil==2.8.1